requirements-test.txt
# Requirements for executing tests in test environments
coverage
mock>=1.0.1
nose>=1.3.0
django-nose>=1.2

# Required to test default models
django-extensions>=0.9
django-mptt
django-mptt-admin
